requirements.txt
Flask==2.3.2
pandas
google-generativeai==0.3.2
requests
python-dotenv==1.0.0
